requirements.txt
h5py==3.6.0
matplotlib==3.5.1
neptune_client==0.9.16
numpy==1.22.2
pandas==1.4.0
pytorch_lightning==1.3.5
scikit_learn==1.2.1
scipy==1.6.2
torch==1.10.2
